# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Reputation & Visibility Audit API",
        "description": "A comprehensive API for auditing business online presence, SEO, social media, and Google Business Profile. Provides AI-powered recommendations via GPT-4.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/subscription/current": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get current subscription",
                "description": "Retrieve user's active subscription",
                "operationId": "getCurrentSubscription",
                "responses": {
                    "200": {
                        "description": "Subscription retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/subscription/upgrade": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Upgrade subscription",
                "description": "Upgrade to a higher tier plan",
                "operationId": "upgradeSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subscription_plan_id"
                                ],
                                "properties": {
                                    "subscription_plan_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "billing_cycle": {
                                        "type": "string",
                                        "enum": [
                                            "monthly",
                                            "annual"
                                        ],
                                        "example": "monthly"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription upgraded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/subscription/cancel": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Cancel subscription",
                "description": "Cancel the active subscription",
                "operationId": "cancelSubscription",
                "responses": {
                    "200": {
                        "description": "Subscription cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription canceled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/subscription/usage": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get subscription usage",
                "description": "Get current usage statistics against plan limits",
                "operationId": "getSubscriptionUsage",
                "responses": {
                    "200": {
                        "description": "Usage retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "audits_used": {
                                                    "type": "integer"
                                                },
                                                "audits_limit": {
                                                    "type": "integer"
                                                },
                                                "audits_percent": {
                                                    "type": "integer"
                                                },
                                                "businesses_used": {
                                                    "type": "integer"
                                                },
                                                "businesses_limit": {
                                                    "type": "integer"
                                                },
                                                "businesses_percent": {
                                                    "type": "integer"
                                                },
                                                "period_start": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "period_end": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
...
}